+++ traveller/src/main/java/kr/event/action/EventUpdateAction.java
@@ -22,9 +22,8 @@ public class EventUpdateAction implements Action {
 		Integer user_num = (Integer)session.getAttribute("user_num");
 		int user_auth = (Integer)session.getAttribute("user_auth");
 		
-		if(user_num==null || user_auth!=9) {
-			return "redirect:/main/main.do";
-		}
+		//로그인, 관리자 조건체크
+		if(user_num==null || user_auth!=9) return "redirect:/main/main.do";
 
 		MultipartRequest multi = FileUtil.createFile(request);
 		
@@ -32,25 +31,23 @@ public class EventUpdateAction implements Action {
 		int event_num = Integer.parseInt(multi.getParameter("event_num"));
 		
 		EventDAO dao = EventDAO.getInstance();
-		EventVO db_event = dao.getEventDetail(event_num);
+		EventVO db_event = dao.getEventDetail(event_num); //수정 전 데이터
 		
 		SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
 		SimpleDateFormat newformat = new SimpleDateFormat("yy/MM/dd");
 		
+		//수정할 데이터 저장
 		EventVO event = new EventVO();
 		event.setEvent_num(event_num);
 		event.setEvent_start(newformat.format(format.parse(multi.getParameter("event_start"))));
 		event.setEvent_end(newformat.format(format.parse(multi.getParameter("event_end"))));
-		if(event_photo1!=null) {
-			event.setEvent_photo1(event_photo1);
-		} else {
-			event.setEvent_photo1(db_event.getEvent_photo1());
-			FileUtil.removeFile(request, db_event.getEvent_photo1());
-		}
+		event.setEvent_photo1(event_photo1);
 		event.setEvent_content(multi.getParameter("event_content"));
 		event.setEvent_count(Integer.parseInt(multi.getParameter("event_count")));
 		
 		dao.eventUpdate(event);
+		
+		if(event_photo1!=null) FileUtil.removeFile(request, db_event.getEvent_photo1()); //기존 파일 삭제
 		
 		request.setAttribute("notice_msg","이벤트를 수정했습니다!");
 		request.setAttribute("notice_url","eventDetail.do?event_num="+event_num);
